+++ src/controllers/EntriesController.php
@@ -95,7 +95,11 @@ class EntriesController extends BaseEntriesController
         $entry = Craft::createObject(Entry::class);
         $entry->siteId = $site->id;
         $entry->sectionId = $section->id;
-        $entry->authorId = $this->request->getParam('authorId', $user->id);
+
+        $authorIds = $this->request->getQueryParam('authorIds') ?? $this->request->getQueryParam('authorId');
+        if ($authorIds !== null) {
+            $entry->setAuthorIds($authorIds);
+        }
 
         // Type
         if (($typeHandle = $this->request->getParam('type')) !== null) {
@@ -239,7 +243,7 @@ class EntriesController extends BaseEntriesController
         if (
             $entry->id &&
             !$duplicate &&
-            $entry->authorId != $currentUser->id &&
+            !in_array($currentUser->id, $entry->getAuthorIds(), true) &&
             $section->type !== Section::TYPE_SINGLE &&
             $entry->enabled
         ) {
@@ -465,14 +469,13 @@ class EntriesController extends BaseEntriesController
         $fieldsLocation = $this->request->getParam('fieldsLocation', 'fields');
         $entry->setFieldValuesFromRequest($fieldsLocation);
 
-        // Author
-        $authorId = $this->request->getBodyParam('author', ($entry->authorId ?: static::currentUser()->id));
-
-        if (is_array($authorId)) {
-            $authorId = $authorId[0] ?? null;
+        // Authors
+        $authorIds = $this->request->getBodyParam('authors') ?? $this->request->getBodyParam('author');
+        if ($authorIds !== null) {
+            $entry->setAuthorIds($authorIds);
+        } elseif (!$entry->id) {
+            $entry->setAuthor(static::currentUser());
         }
-
-        $entry->authorId = $authorId;
 
         // Parent
         if (($parentId = $this->request->getBodyParam('parentId')) !== null) {