+++ src/elements/User.php
@@ -2312,9 +2312,9 @@ JS, [
 
                 // Update the entry/version/draft tables to point to the new user
                 $userRefs = [
-                    Table::ENTRIES => 'authorId',
                     Table::DRAFTS => 'creatorId',
                     Table::REVISIONS => 'creatorId',
+                    Table::ENTRIES_AUTHORS => 'authorId',
                 ];
 
                 foreach ($userRefs as $table => $column) {
@@ -2334,7 +2334,10 @@ JS, [
 
                 foreach (Db::each($entryQuery) as $entry) {
                     /** @var Entry $entry */
-                    $elementsService->deleteElement($entry);
+                    // only delete their entry if they're the sole author
+                    if ($entry->getAuthorIds() === [$this->id]) {
+                        $elementsService->deleteElement($entry);
+                    }
                 }
             }