+++ src/models/Volume.php
@@ -200,22 +200,26 @@ class Volume extends Model implements BaseFsInterface, FieldLayoutProviderInterf
         ];
         $rules[] = [['fieldLayout'], 'validateFieldLayout'];
         $rules[] = [['subpath'], fn($attribute) => $this->validateUniqueSubpath($attribute), 'skipOnEmpty' => false];
-        $rules[] = [
-            ['fsHandle'],
-            'compare',
-            'compareAttribute' => 'fsHandle',
-            'compareValue' => App::parseEnv(Craft::$app->getConfig()->getGeneral()->tempAssetUploadFs),
-            'operator' => '!=',
-            'message' => Craft::t('app', 'This filesystem has been reserved for temporary asset uploads. Please choose a different one for your volume.'),
-        ];
-        $rules[] = [
-            ['transformFsHandle'],
-            'compare',
-            'compareAttribute' => 'transformFsHandle',
-            'compareValue' => App::parseEnv(Craft::$app->getConfig()->getGeneral()->tempAssetUploadFs),
-            'operator' => '!=',
-            'message' => Craft::t('app', 'This filesystem has been reserved for temporary asset uploads. Please choose a different one for your volume.'),
-        ];
+
+        $tempAssetUploadFs = App::parseEnv(Craft::$app->getConfig()->getGeneral()->tempAssetUploadFs);
+        if ($tempAssetUploadFs) {
+            $rules[] = [
+                ['fsHandle'],
+                'compare',
+                'compareAttribute' => 'fsHandle',
+                'compareValue' => $tempAssetUploadFs,
+                'operator' => '!=',
+                'message' => Craft::t('app', 'This filesystem has been reserved for temporary asset uploads. Please choose a different one for your volume.'),
+            ];
+            $rules[] = [
+                ['transformFsHandle'],
+                'compare',
+                'compareAttribute' => 'transformFsHandle',
+                'compareValue' => $tempAssetUploadFs,
+                'operator' => '!=',
+                'message' => Craft::t('app', 'This filesystem has been reserved for temporary asset uploads. Please choose a different one for your volume.'),
+            ];
+        }
 
         return $rules;
     }