+++ src/web/assets/cp/src/js/ElementEditor.js
@@ -11,6 +11,7 @@ Craft.ElementEditor = Garnish.Base.extend(
     $activityContainer: null,
     $tabContainer: null,
     $contentContainer: null,
+    $sidebar: null,
     $spinner: null,
     $expandSiteStatusesBtn: null,
     $statusIcon: null,
@@ -83,9 +84,11 @@ Craft.ElementEditor = Garnish.Base.extend(
       if (this.isFullPage) {
         this.$tabContainer = $('#tabs');
         this.$contentContainer = $('#content');
+        this.$sidebar = $('#details .details');
       } else {
         this.$tabContainer = this.slideout.$tabContainer;
         this.$contentContainer = this.slideout.$content;
+        this.$sidebar = this.slideout.$sidebar;
       }
 
       this.queue = this._createQueue();
@@ -1376,16 +1379,24 @@ Craft.ElementEditor = Garnish.Base.extend(
             }
 
             // Add missing field modified indicators
-            const selectors = response.data.modifiedAttributes
+            const selector = response.data.modifiedAttributes
               .map((attr) => {
                 attr = this.namespaceInputName(attr);
-                return `[name="${attr}"],[name^="${attr}["]`;
+                return [`[name="${attr}"]`, `[name^="${attr}["]`];
               })
-              .concat(modifiedFieldNames.map((name) => `[name="${name}"]`));
+              .flat()
+              .concat(modifiedFieldNames.map((name) => `[name="${name}"]`))
+              .join(',');
 
-            let $fields = $(selectors.join(','))
+            let $fields = this.$contentContainer
+              .find(selector)
               .parents()
-              .filter('.flex-fields > .field:not(:has(> .status-badge))');
+              .filter('.flex-fields > .field:not(:has(> .status-badge))')
+              .add(
+                this.$sidebar
+                  .find(selector)
+                  .closest('.field:not(:has(> .status-badge))')
+              );
 
             if (params.dirtyFields) {
               $fields = $fields.add(